refactor!: KeyPair derive_wallet, add tests (#38)
* refactor!: remove the old derive_wallet function on KeyPair

Replace the old derive_wallet function with a version that does not
take the blockchain_client as an argument.

The old derive_wallet function was requiring the blockchain_client as an argument.
However, providing the derive_wallet function with no additional arguments
other than the &self Keypair struct should be sufficient
as the blockchain client can be sent on the result
from derive_wallet with other methods.
This requires anything that called the derive_wallet function to be refactored
to not pass the blockchain client as a argument.
Instead to specify the blockchain client, can use the set_blockchain_client
function on the wallet.

* refactor: refactor workflow example to use new version of derive wallet, use walletd::Error in workflow example - change unwrap to ?,

* refactor!: remove pub for KeyPair fields

BREAKING CHANGE: Change use of derive_wallet function to not pass the blockchain connector as an argument.
No longer allowed to access fields of the KeyPair struct directly or to modify directly. Can still use the public getter functions on KeyPair to get the values of the fields.